+++ Event/FeatureTypeEvent.php
@@ -20,18 +20,18 @@ use Thelia\Model\Feature;
 /**
  * Class FeatureTypeEvent
  * @package FeatureType\Event
- * @author Gilles Bourgeat <gbourgeat@openstudio.fr>
+ * @author Gilles Bourgeat <gilles.bourgeat@gmail.com>
  */
 class FeatureTypeEvent extends ActionEvent
 {
     /** @var ConnectionInterface|null */
-    private $connectionInterface = null;
+    protected $connectionInterface = null;
 
     /** @var FeatureType */
-    private $featureType = null;
+    protected $featureType = null;
 
     /** @var Feature|null */
-    private $feature = null;
+    protected $feature = null;
 
     /**
      * @param FeatureType $featureType