+++ Loop/FeatureAvailabilityExtendLoop.php
@@ -30,7 +30,7 @@ use Thelia\Model\Map\FeatureAvTableMap;
 /**
  * Class FeatureAvailabilityExtendLoop
  * @package FeatureType\Loop
- * @author Gilles Bourgeat <gbourgeat@openstudio.fr>
+ * @author Gilles Bourgeat <gilles.bourgeat@gmail.com>
  */
 class FeatureAvailabilityExtendLoop extends FeatureAvailability implements PropelSearchLoopInterface
 {
@@ -125,7 +125,7 @@ class FeatureAvailabilityExtendLoop extends FeatureAvailability implements Prope
      * @param LoopResult $loopResult
      * @return array|mixed|\Propel\Runtime\Collection\ObjectCollection
      */
-    private function getFeaturesMeta(LoopResult $loopResult)
+    protected function getFeaturesMeta(LoopResult $loopResult)
     {
         $featureAvIds = array();
 
@@ -175,7 +175,7 @@ class FeatureAvailabilityExtendLoop extends FeatureAvailability implements Prope
      * @param string $slug
      * @return string
      */
-    private function formatSlug($slug)
+    protected function formatSlug($slug)
     {
         return strtoupper(str_replace('-', '_', $slug));
     }