+++ Loop/FeatureExtendLoop.php
@@ -30,7 +30,7 @@ use Thelia\Model\Map\FeatureTableMap;
 /**
  * Class FeatureExtendLoop
  * @package FeatureType\Loop
- * @author Gilles Bourgeat <gbourgeat@openstudio.fr>
+ * @author Gilles Bourgeat <gilles.bourgeat@gmail.com>
  */
 class FeatureExtendLoop extends Feature implements PropelSearchLoopInterface
 {
@@ -125,7 +125,7 @@ class FeatureExtendLoop extends Feature implements PropelSearchLoopInterface
      * @param LoopResult $loopResult
      * @return array|mixed|\Propel\Runtime\Collection\ObjectCollection
      */
-    private function getFeaturesType(LoopResult $loopResult)
+    protected function getFeaturesType(LoopResult $loopResult)
     {
         $featureIds = array();
 
@@ -160,7 +160,7 @@ class FeatureExtendLoop extends Feature implements PropelSearchLoopInterface
      * @param string $slug
      * @return string
      */
-    private function formatSlug($slug)
+    protected function formatSlug($slug)
     {
         return strtoupper(str_replace('-', '_', $slug));
     }