+++ Loop/FeatureValueExtendLoop.php
@@ -29,7 +29,7 @@ use Thelia\Model\Map\FeatureAvTableMap;
 /**
  * Class FeatureValueExtendLoop
  * @package FeatureType\Loop
- * @author Gilles Bourgeat <gbourgeat@openstudio.fr>
+ * @author Gilles Bourgeat <gilles.bourgeat@gmail.com>
  */
 class FeatureValueExtendLoop extends FeatureValue implements PropelSearchLoopInterface
 {
@@ -37,7 +37,7 @@ class FeatureValueExtendLoop extends FeatureValue implements PropelSearchLoopInt
      * @param LoopResult $loopResult
      * @return array|mixed|\Propel\Runtime\Collection\ObjectCollection
      */
-    private function getFeaturesMeta(LoopResult $loopResult)
+    protected function getFeaturesMeta(LoopResult $loopResult)
     {
         $featureAvIds = array();
 
@@ -87,7 +87,7 @@ class FeatureValueExtendLoop extends FeatureValue implements PropelSearchLoopInt
      * @param string $slug
      * @return string
      */
-    private function formatSlug($slug)
+    protected function formatSlug($slug)
     {
         return strtoupper(str_replace('-', '_', $slug));
     }